source/temp_kinetix_ne/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argparse==1.4.0
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
brax==0.12.1
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
chex==0.1.90
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
craftax==1.5.0
cycler==0.12.1
decorator==4.4.2
distlib==0.4.0
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.9
equinox==0.11.4
etils==1.13.0
executing==2.2.1
farama-notifications==0.0.4
filelock==3.19.1
flask==3.1.2
flask-cors==6.0.1
flax==0.8.5
fonttools==4.59.2
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
gitpython==3.1.45
glfw==2.10.0
grpcio==1.74.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.0
gymnax==0.0.9
humanize==4.13.0
hydra-core==1.3.2
identify==2.6.14
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-resources==6.5.2
ipython==9.5.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jax==0.4.26
jax-cuda12-pjrt==0.4.26
jax-cuda12-plugin==0.4.26
jax2d==1.0.1
jaxgl==1.0.1
jaxlib==0.4.26
jaxopt==0.8.5
jaxtyping==0.3.2
jaxued==0.0.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
-e file:///home/eleni/workspace/continual_control_through_neuroevolution/methods/Kinetix
kiwisolver==1.4.9
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
ml-collections==1.1.0
ml-dtypes==0.5.3
moviepy==1.0.3
msgpack==1.1.1
mujoco==3.2.4
mujoco-mjx==3.2.4
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
omegaconf==2.3.0
opt-einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.6.4
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pre-commit==4.3.0
proglog==0.1.12
prompt-toolkit==3.0.52
protobuf==6.32.1
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.8
pydantic-core==2.33.2
pygame==2.6.1
pygame-widgets==1.1.5
pygments==2.19.2
pyopengl==3.1.10
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytinyrenderer==0.0.14
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.37.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
tensorboardx==2.6.4
tensorflow-probability==0.25.0
tensorstore==0.1.76
threadpoolctl==3.6.0
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
trimesh==4.8.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
wadler-lindig==0.1.7
wandb==0.21.4
wcwidth==0.2.13
werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
